Dto's gefixt

+++ src/App/Domain/Kunstwerk.cs
@@ -9,7 +9,7 @@ namespace Domain
         public int Id { get; set; }
         public string Naam { get; private set; }
         public DateTime Einddatum { get; private set; }
-        public double Prijs { get; private set; }
+        public decimal Prijs { get; private set; }
         public string Beschrijving { get; private set; }
         public List<Foto> Fotos { get; private set; }
         public bool TeKoop { get; set; }
@@ -17,7 +17,7 @@ namespace Domain
         public string Materiaal { get; private set; }
         public string NaamKunstenaar { get; private set; }
 
-        public Kunstwerk(string naam, DateTime einddatum, double prijs, string beschrijving, List<Foto> fotos, bool isVeilbaar, string materiaal, string naamKunstenaar)
+        public Kunstwerk(string naam, DateTime einddatum, decimal prijs, string beschrijving, List<Foto> fotos, bool isVeilbaar, string materiaal, string naamKunstenaar)
         {
             Naam = naam;
             Einddatum = einddatum;
